+++ ground/src/main/java/com/google/android/ground/ui/offlineareas/OfflineAreaListAdapter.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.offlinebasemap
+package com.google.android.ground.ui.offlineareas
 
 import android.view.LayoutInflater
 import android.view.View
