+++ ground/src/main/java/com/google/android/ground/ui/offlineareas/selector/OfflineAreaSelectorViewModel.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.offlinebasemap.selector
+package com.google.android.ground.ui.offlineareas.selector
 
 import android.content.res.Resources
 import androidx.lifecycle.MutableLiveData
