+++ ground/src/main/res/layout/download_progress_dialog_frag.xml
@@ -21,7 +21,7 @@
     <import type="android.view.View" />
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.offlinebasemap.selector.OfflineAreaSelectorViewModel" />
+      type="com.google.android.ground.ui.offlineareas.selector.OfflineAreaSelectorViewModel" />
   </data>
 
   <FrameLayout
