+++ ground/src/main/res/layout/offline_area_selector_frag.xml
@@ -24,7 +24,7 @@
     <import type="android.view.View" />
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.offlinebasemap.selector.OfflineAreaSelectorViewModel" />
+      type="com.google.android.ground.ui.offlineareas.selector.OfflineAreaSelectorViewModel" />
   </data>
 
   <FrameLayout
