+++ ground/src/main/res/layout/offline_area_viewer_frag.xml
@@ -22,7 +22,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.offlinebasemap.viewer.OfflineAreaViewerViewModel" />
+      type="com.google.android.ground.ui.offlineareas.viewer.OfflineAreaViewerViewModel" />
   </data>
 
   <androidx.constraintlayout.widget.ConstraintLayout
