+++ ground/src/main/res/layout/offline_areas_frag.xml
@@ -22,7 +22,7 @@
   <data>
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.offlinebasemap.OfflineAreasViewModel" />
+      type="com.google.android.ground.ui.offlineareas.OfflineAreasViewModel" />
   </data>
 
   <androidx.constraintlayout.widget.ConstraintLayout
